move to micronaut-langchain-store-redis
redis docs
