+++ micronaut-langchain4j-core/build.gradle.kts
@@ -8,7 +8,6 @@ dependencies {
     api(mn.micronaut.context)
     api(libs.langchain4j)
     implementation(platform(libs.langchain4j.community.bom))
-    compileOnly(libs.langchain4j.community.redis)
     compileOnly(libs.langchain4j.community.neo4j)
     compileOnly(libs.langchain4j.cassandra)
     testAnnotationProcessor(mn.micronaut.inject.java)
