+++ micronaut-langchain4j-store-redis/src/test/java/io/micronaut/langchain4j/redis/memory/RedisChatMemoryStoreConfigurationTest.java
@@ -1,13 +1,11 @@
-package example.micronaut;
+package io.micronaut.langchain4j.redis.memory;
 
 import io.micronaut.context.annotation.Property;
 import io.micronaut.core.util.StringUtils;
-import io.micronaut.langchain4j.store.memory.chat.redis.RedisChatMemoryStoreConfiguration;
 import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Property(name = "langchain4j.ollama.enabled", value = StringUtils.FALSE)
